+++ src/trace_processor/perfetto_sql/stdlib/viz/summary/trace.sql
@@ -14,7 +14,7 @@
 -- limitations under the License.
 
 -- Returns 1 if trace is linux trace and 0 otherwise.
-CREATE PERFETTO FUNCTION _is_linux_trace() RETURNS INT AS
+CREATE PERFETTO FUNCTION _is_linux_trace() RETURNS LONG AS
 SELECT count() == 1
 FROM metadata
 WHERE name = 'system_name' AND str_value = 'Linux';
